chore(sample): SampleApp renamed to SampleStores
