+++ Samples/SampleStores/Commands/CountryCommand.cs
@@ -2,10 +2,10 @@
 using System.CommandLine.Invocation;
 using System.CommandLine.Rendering;
 using JsonStores.Exceptions;
-using JsonStores.Samples.SampleApp.Models;
+using JsonStores.Samples.SampleStores.Models;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace JsonStores.Samples.SampleApp.Commands
+namespace JsonStores.Samples.SampleStores.Commands
 {
     public static class CountryCommand
     {
