+++ Samples/SampleStores/Models/Country.cs
@@ -1,6 +1,6 @@
 ﻿using JsonStores.Annotations;
 
-namespace JsonStores.Samples.SampleApp.Models
+namespace JsonStores.Samples.SampleStores.Models
 {
     public class Country
     {
