+++ Samples/SampleStores/Models/Note.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace JsonStores.Samples.SampleApp.Models
+namespace JsonStores.Samples.SampleStores.Models
 {
     public class Note
     {
