+++ Samples/SampleStores/README.md
@@ -1,7 +1,7 @@
-# JSON Stores Sample App
+# Sample Stores App
 
 Basic usage using [System.CommandLine](https://github.com/dotnet/command-line-api) API
 
 ## Instructions
 
-Use `sample --help` to get the availabe commands
+Use `samples --help` to get the available commands
